+++ example/ios/Runner/AppDelegate.m
@@ -15,7 +15,7 @@
     [ThrioModule init:[MainModule new]];
 //    [ThrioModule init:[MainModule new] multiEngineEnabled:YES];
 
-    UINavigationController *nvc = [[NavigatorNavigationController alloc] initWithUrl:@"/biz1/native1" params:nil];
+    UINavigationController *nvc = [[NavigatorNavigationController alloc] initWithUrl:@"/biz1/flutter1" params:nil];
     self.window.rootViewController = nvc;
     [self.window makeKeyAndVisible];