+++ example/lib/src/biz2/flutter4/page.dart
@@ -30,7 +30,7 @@ class _PageState extends State<Page> {
         backgroundColor: Colors.white,
         title:
             const Text('thrio_example', style: TextStyle(color: Colors.black)),
-        leading: context.canPop(const IconButton(
+        leading: context.shouldCanPop(const IconButton(
           color: Colors.black,
           tooltip: 'back',
           icon: Icon(Icons.arrow_back_ios),